+++ crates/assembly-syntax/src/sema/passes/mod.rs
@@ -1,6 +1,10 @@
 mod const_eval;
+mod verify_empty_control_flow;
 mod verify_invoke;
 mod verify_repeat;
 
 pub(super) use self::verify_invoke::{LocalInvokeTarget, VerifyInvokeTargets};
-pub use self::{const_eval::ConstEvalVisitor, verify_repeat::VerifyRepeatCounts};
+pub use self::{
+    const_eval::ConstEvalVisitor, verify_empty_control_flow::VerifyEmptyControlFlow,
+    verify_repeat::VerifyRepeatCounts,
+};